fixed broken merge stuff & threw nicolai out the window
